+++ examples/by_project/unixstl/filesystem/readdir_sequence/main.cpp
@@ -57,90 +57,160 @@ int main(int argc, char* argv[])
 
 
     {
+        std::cout << std::endl;
         std::cout << "searching for directories in '" << root_dir << "':" << std::endl;
 
-        readdir_sequence directories(root_dir, readdir_sequence::absolutePath | readdir_sequence::fullPath | readdir_sequence::directories);
-
-        for (readdir_sequence::const_iterator i = directories.begin(); directories.end() != i; ++i)
+        try
         {
-            std::cout
-                << "\t"
-                << *i
-                << "/"
-                << std::endl;
+            readdir_sequence directories(root_dir, readdir_sequence::absolutePath | readdir_sequence::fullPath | readdir_sequence::directories);
+
+            for (readdir_sequence::const_iterator i = directories.begin(); directories.end() != i; ++i)
+            {
+                std::cout
+                    << "\t"
+                    << *i
+                    << "/"
+                    << std::endl;
+            }
+        }
+        catch (unixstl::unixstl_exception& x)
+        {
+            std::cerr << "EXCEPTION: " << x.what() << std::endl;
         }
     }
 
     {
+        std::cout << std::endl;
         std::cout << "searching for files in '" << root_dir << "':" << std::endl;
 
-        readdir_sequence files(root_dir, readdir_sequence::absolutePath | readdir_sequence::fullPath | readdir_sequence::files);
-
-        for (readdir_sequence::const_iterator i = files.begin(); files.end() != i; ++i)
+        try
         {
-            std::cout
-                << "\t"
-                << *i
-                << std::endl;
+            readdir_sequence files(root_dir, readdir_sequence::absolutePath | readdir_sequence::fullPath | readdir_sequence::files);
+
+            for (readdir_sequence::const_iterator i = files.begin(); files.end() != i; ++i)
+            {
+                std::cout
+                    << "\t"
+                    << *i
+                    << std::endl;
+            }
+        }
+        catch (unixstl::unixstl_exception& x)
+        {
+            std::cerr << "EXCEPTION: " << x.what() << std::endl;
         }
     }
 
     {
+        std::cout << std::endl;
+        std::cout << "searching for devices in '" << root_dir << "':" << std::endl;
+
+        try
+        {
+            readdir_sequence devices(root_dir, readdir_sequence::absolutePath | readdir_sequence::fullPath | readdir_sequence::devices);
+
+            for (readdir_sequence::const_iterator i = devices.begin(); devices.end() != i; ++i)
+            {
+                std::cout
+                    << "\t"
+                    << *i
+                    << std::endl;
+            }
+        }
+        catch (unixstl::unixstl_exception& x)
+        {
+            std::cerr << "EXCEPTION: " << x.what() << std::endl;
+        }
+    }
+
+    {
+        std::cout << std::endl;
         std::cout << "searching for sockets in '" << root_dir << "':" << std::endl;
 
-        readdir_sequence sockets(root_dir, readdir_sequence::absolutePath | readdir_sequence::fullPath | readdir_sequence::sockets);
-
-        for (readdir_sequence::const_iterator i = sockets.begin(); sockets.end() != i; ++i)
+        try
         {
-            std::cout
-                << "\t"
-                << *i
-                << std::endl;
+            readdir_sequence sockets(root_dir, readdir_sequence::absolutePath | readdir_sequence::fullPath | readdir_sequence::sockets);
+
+            for (readdir_sequence::const_iterator i = sockets.begin(); sockets.end() != i; ++i)
+            {
+                std::cout
+                    << "\t"
+                    << *i
+                    << std::endl;
+            }
+        }
+        catch (unixstl::unixstl_exception& x)
+        {
+            std::cerr << "EXCEPTION: " << x.what() << std::endl;
         }
     }
 
     {
+        std::cout << std::endl;
         std::cout << "searching for files + sockets in '" << root_dir << "':" << std::endl;
 
-        readdir_sequence entries(root_dir, readdir_sequence::absolutePath | readdir_sequence::fullPath | readdir_sequence::files | readdir_sequence::sockets);
-
-        for (readdir_sequence::const_iterator i = entries.begin(); entries.end() != i; ++i)
+        try
         {
-            std::cout
-                << "\t"
-                << *i
-                << (fs_traits_t::is_directory(*i) ? "/" : "")
-                << std::endl;
+            readdir_sequence entries(root_dir, readdir_sequence::absolutePath | readdir_sequence::fullPath | readdir_sequence::files | readdir_sequence::sockets);
+
+            for (readdir_sequence::const_iterator i = entries.begin(); entries.end() != i; ++i)
+            {
+                std::cout
+                    << "\t"
+                    << *i
+                    << (fs_traits_t::is_directory(*i) ? "/" : "")
+                    << std::endl;
+            }
+        }
+        catch (unixstl::unixstl_exception& x)
+        {
+            std::cerr << "EXCEPTION: " << x.what() << std::endl;
         }
     }
 
     {
+        std::cout << std::endl;
         std::cout << "searching for directories + files + sockets in '" << root_dir << "':" << std::endl;
 
-        readdir_sequence entries(root_dir, readdir_sequence::absolutePath | readdir_sequence::fullPath | readdir_sequence::directories | readdir_sequence::files | readdir_sequence::sockets);
-
-        for (readdir_sequence::const_iterator i = entries.begin(); entries.end() != i; ++i)
+        try
         {
-            std::cout
-                << "\t"
-                << *i
-                << (fs_traits_t::is_directory(*i) ? "/" : "")
-                << std::endl;
+            readdir_sequence entries(root_dir, readdir_sequence::absolutePath | readdir_sequence::fullPath | readdir_sequence::directories | readdir_sequence::files | readdir_sequence::sockets);
+
+            for (readdir_sequence::const_iterator i = entries.begin(); entries.end() != i; ++i)
+            {
+                std::cout
+                    << "\t"
+                    << *i
+                    << (fs_traits_t::is_directory(*i) ? "/" : "")
+                    << std::endl;
+            }
+        }
+        catch (unixstl::unixstl_exception& x)
+        {
+            std::cerr << "EXCEPTION: " << x.what() << std::endl;
         }
     }
 
     {
+        std::cout << std::endl;
         std::cout << "searching for all types in '" << root_dir << "':" << std::endl;
 
-        readdir_sequence entries(root_dir, readdir_sequence::absolutePath | readdir_sequence::fullPath);
-
-        for (readdir_sequence::const_iterator i = entries.begin(); entries.end() != i; ++i)
+        try
         {
-            std::cout
-                << "\t"
-                << *i
-                << (fs_traits_t::is_directory(*i) ? "/" : "")
-                << std::endl;
+            readdir_sequence entries(root_dir, readdir_sequence::absolutePath | readdir_sequence::fullPath);
+
+            for (readdir_sequence::const_iterator i = entries.begin(); entries.end() != i; ++i)
+            {
+                std::cout
+                    << "\t"
+                    << *i
+                    << (fs_traits_t::is_directory(*i) ? "/" : "")
+                    << std::endl;
+            }
+        }
+        catch (unixstl::unixstl_exception& x)
+        {
+            std::cerr << "EXCEPTION: " << x.what() << std::endl;
         }
     }