+++ include/winstl/filesystem/findfile_sequence.hpp
@@ -18,7 +18,7 @@
  *          ownership issues described in the article.
  *
  * Created: 15th January 2002
- * Updated: 17th April 2025
+ * Updated: 30th April 2025
  *
  * Thanks:  To Nevin Liber for pressing upon me the need to lead by example
  *          when writing books about good design/implementation; to Florin L
@@ -71,9 +71,9 @@
 
 #ifndef STLSOFT_DOCUMENTATION_SKIP_SECTION
 # define WINSTL_VER_WINSTL_FILESYSTEM_HPP_FINDFILE_SEQUENCE_MAJOR       4
-# define WINSTL_VER_WINSTL_FILESYSTEM_HPP_FINDFILE_SEQUENCE_MINOR       11
+# define WINSTL_VER_WINSTL_FILESYSTEM_HPP_FINDFILE_SEQUENCE_MINOR       12
 # define WINSTL_VER_WINSTL_FILESYSTEM_HPP_FINDFILE_SEQUENCE_REVISION    0
-# define WINSTL_VER_WINSTL_FILESYSTEM_HPP_FINDFILE_SEQUENCE_EDIT        276
+# define WINSTL_VER_WINSTL_FILESYSTEM_HPP_FINDFILE_SEQUENCE_EDIT        277
 #endif /* !STLSOFT_DOCUMENTATION_SKIP_SECTION */
 
 
@@ -280,6 +280,7 @@ public:
         ,   directories                 =   0x0010  //!< Causes the search to include directories
         ,   files                       =   0x0020  //!< Causes the search to include files
 #ifndef STLSOFT_DOCUMENTATION_SKIP_SECTION
+        ,   devices                     =   0x0000  /*!< This has no effect on Windows, but is here for structural conformance compatibility with unixstl::glob_sequence. */
         ,   sockets                     =   0x0000  /*!< This has no effect on Windows, but is here for structural conformance compatibility with unixstl::glob_sequence. */
 #endif /* !STLSOFT_DOCUMENTATION_SKIP_SECTION */
         ,   skipReparseDirs             =   0x0100  //!< Causes the search to skip directories that are reparse points