+++ HISTORY.md
@@ -4,6 +4,12 @@
 ----
 
 
+30th April 2025 - 1.11.1-beta7 released
+---------------------------------------
+
+ * Support for searching of devices on the file system, primarily concerning `unixstl::glob_sequence` and `unixstl::readdir_sequence`;
+
+
 28th April 2025 - 1.11.1-beta6 released
 ---------------------------------------